Configure iex console to connect to ttyGS0
ttyGS0 is the gadget serial device, which allows you to connect to iex
from a host connected to the OTG USB port on the zero. For example, on
a Mac, executing `screen /dev/tty.usbmodem* 115200` after the zero boots
(approximately 10 seconds for this app) should drop you right into an
iex session. Additionally, Ctrl+C twice will drop to sh.